att Transacoes Service e test
mudanças para funcionar o service e o teste

+++ demo/src/main/java/com/example/demo/enums/TipoDeTransacoesEnums.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 @Getter
 public enum TipoDeTransacoesEnums {
 	
-	DEPOSITO(1, "Deposito"), PAGAMENTO(2, "Pagamento"), TRANSFERENCIA(3,"Transferencia"), SAQUE(4, "Saque");
+	DEPOSITO(1L, "Deposito"), PAGAMENTO(2L, "Pagamento"), TRANSFERENCIA(3L,"Transferencia"), SAQUE(4L, "Saque");
 	
 
 
-	public static TipoDeTransacoesEnums get(Integer id) {
+	public static TipoDeTransacoesEnums get(Long id) {
 		for (TipoDeTransacoesEnums obj : TipoDeTransacoesEnums.values()) {
 			if (obj.getId().equals(id))
 				return obj;
@@ -17,10 +17,10 @@ public enum TipoDeTransacoesEnums {
 		return null;
 	}
 
-	Integer id;
+	Long id;
 	String nome;
 
-	TipoDeTransacoesEnums(Integer id, String nome) {
+	TipoDeTransacoesEnums(Long id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}
